+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.4] - 2021-10-18
+
+### Fixed
+
+- Kill all players if all game ends with a draw to prevent plugin crash.
+- Fix plugin crash when a player kills other player.
+
 ## [0.1.3] - 2021-10-17
 
 ### Added